Create a plugin for the sumenko extension (#1517)
* Create a plugin for the sumenko extension
To allow intellisense between different modules. This is because the imports we use, makes the intellisense unable to locate the files. Both the Module prefix in addition to the different naming scheme.
This plugin under the hood annotates all imports with `---module '...'` where `...` is a findable name for the engine.

* Code Style

* Update comment