Allow places to autofill from PrizePool (#1857)
* Allow places to autofill from PrizePool

Currently a lot of prizepools which call the legacy modules which have many teams in 1 place slot. This allows the PrizePool module to autoassign any slot with the correct sequential order

* Update prize_pool_legacy_custom.lua

Didnt realise that the dependency call would create an object that is permanent. And therefore I can set a true/false in the header call

* Update prize_pool_legacy_custom.lua

* Update prize_pool_legacy_custom.lua